requirements.txt

fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
jinja2>=3.0.0
# langchain # 或您偏好的LLM交互库
# openai # 如果使用OpenAI
python-dotenv>=1.0.0

# 数据分析与可视化
pandas>=2.0.0
matplotlib>=3.0.0
seaborn>=0.12.0

# 向量存储
faiss-cpu # 或 faiss-gpu (如果支持CUDA)

# 图数据库
neo4j>=5.0.0 # Neo4j官方驱动
networkx>=3.0

# Web搜索 (示例，取决于使用的服务)
# google-api-python-client
# beautifulsoup4
# requests

# MAS / 通信 (如果使用特定库)
# asyncio # 内置

# 其他特定依赖
# ...
